+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonManager.cs
@@ -35,7 +35,7 @@ namespace Krypton.Toolkit
         private static PaletteProfessionalOffice2003 _paletteProfessionalOffice2003;
         private static PaletteProfessionalSystem _paletteProfessionalSystem;
 
-        //private static PaletteBase _customPalette;
+        private static IPalette _customPalette;
 
         #region Office 2007 Themes
 
@@ -210,6 +210,8 @@ namespace Krypton.Toolkit
             ResetGlobalApplyToolstrips();
             ResetGlobalAllowFormChrome();
             ResetGlobalStrings();
+
+            _customPalette = null;
         }
 
         /// <summary>
@@ -392,6 +394,14 @@ namespace Krypton.Toolkit
         [Description(@"")]
         [DefaultValue(null)]
         public KryptonCustomPaletteManager CustomPaletteManager { get => _customPaletteManager; set => _customPaletteManager = value; }
+
+        /// <summary>Specify a custom palette outside the existing palettes.</summary>
+        /// <value>A custom palette.</value>
+        [Category(@"Visuals")]
+        [Description(@"Specify a custom palette outside the existing palettes.")]
+        [DefaultValue(null)]
+        public IPalette CustomPalette { get => _customPalette; set => _customPalette = value; }
+
         #endregion
 
         #region Static ApplyToolstrips
